requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-chroma
pypdf